Move DeviceContext from BaseAdresseNationale to MyApp
