+++ pages/_app.js
@@ -1,11 +1,22 @@
-import React, {useEffect} from 'react'
+import React, {useEffect, createContext, useState} from 'react'
 import PropTypes from 'prop-types'
 import Head from 'next/head'
 
 const PIWIK_URL = process.env.NEXT_PUBLIC_PIWIK_URL
 const PIWIK_SITE_ID = process.env.NEXT_PUBLIC_PIWIK_SITE_ID
+const MOBILE_WIDTH = 900
 
-function MyApp({Component, pageProps}) {
+export const DeviceContext = createContext()
+
+function MyApp({Component, pageProps, isSafariBrowser}) {
+  const [viewHeight, setViewHeight] = useState('100vh')
+  const [isMobileDevice, setIsMobileDevice] = useState(false)
+
+  const handleResize = () => {
+    setViewHeight(`${window.innerWidth}px`)
+    setIsMobileDevice(window.innerWidth < MOBILE_WIDTH)
+  }
+
   const logPageView = () => {
     if (window.Piwik) {
       const tracker = window.Piwik.getTracker(`${PIWIK_URL}/piwik.php`, PIWIK_SITE_ID)
@@ -22,13 +33,28 @@ function MyApp({Component, pageProps}) {
     }, 400)
   })
 
+  useEffect(() => {
+    window.addEventListener('resize', handleResize)
+    handleResize()
+
+    return () => {
+      window.removeEventListener('resize', handleResize)
+    }
+  }, [])
+
   return (
     <>
       <Head>
         <meta name='viewport' content='width=device-width, initial-scale=1' />
       </Head>
-      <Component {...pageProps} />
-
+      <DeviceContext.Provider value={{
+        viewHeight,
+        isMobileDevice,
+        isSafariBrowser
+      }}
+      >
+        <Component {...pageProps} />
+      </DeviceContext.Provider>
       <style global jsx>{`
         body,
         html,
@@ -47,7 +73,17 @@ function MyApp({Component, pageProps}) {
 
 MyApp.propTypes = {
   Component: PropTypes.func.isRequired,
-  pageProps: PropTypes.object.isRequired
+  pageProps: PropTypes.object.isRequired,
+  isSafariBrowser: PropTypes.bool.isRequired
+}
+
+MyApp.getInitialProps = async ({ctx}) => {
+  const {req} = ctx
+  const {headers} = req
+  const userAgent = headers['user-agent']
+  const isSafariBrowser = userAgent.toLowerCase().includes('safari')
+
+  return {isSafariBrowser}
 }
 
 export default MyApp
